+++ .line-cap-allowlist.tsv
@@ -21,8 +21,7 @@ crates/trusty-analyze/src/lang/adapters/rust.rs	530
 crates/trusty-analyze/src/lang/adapters/scala.rs	583
 crates/trusty-analyze/src/lang/adapters/swift.rs	531
 crates/trusty-analyze/src/lang/adapters/typescript.rs	680
-crates/trusty-analyze/src/lang/detection.rs	550
-crates/trusty-analyze/src/main.rs	870
+crates/trusty-analyze/src/main.rs	849
 crates/trusty-analyze/src/mcp/mod.rs	1178
 crates/trusty-analyze/tests/stdio_harness.rs	639
 crates/trusty-common/src/bm25_client.rs	503
